Create separate component for pages
