+++ src/app/(main)/layout.tsx
@@ -1,7 +1,8 @@
 "use client";
+import { useEffect } from "react";
+
 import Footer from "@/components/Footer";
 import Header from "@/components/Header";
-import { useEffect } from "react";
 
 export default function RootLayout({
 	children,
