+++ src/app/(main)/pre-order/components/MemberCard.tsx
@@ -1,32 +1,13 @@
+import type { IMemberCardModel } from "@/utils/models/IMemberCardModel";
 import Image from "next/image";
 
-export default function MemberCard({
-	name,
-	price,
-	doseTitle,
-	doseValue,
-	discountTitle,
-	capsulePrice,
-	spotsRemainds,
-	forever,
-	isActive,
-}: {
-	name: string;
-	doseTitle: string;
-	discountTitle?: string;
-	doseValue?: string;
-	capsulePrice?: number;
-	spotsRemainds?: number;
-	price: number;
-	forever?: boolean;
-	isActive?: boolean;
-}) {
+export default function MemberCard(model: Readonly<IMemberCardModel>) {
 	return (
 		<div
 			className={`grid gap-[16px] p-[16px] md:px-[36px] items-center rounded-[8px] mx-auto  w-full
-                ${isActive ? "grid-cols-[52px_1fr_auto] bg-white shadow-3 h-[138px] mx-[-16px] md:mx-[0]" : "bg-grey21 h-[106px] md:w-[calc(100%-72px)]"}`}
+                ${model.isActive ? "grid-cols-[52px_1fr_auto] bg-white shadow-3 h-[138px] mx-[-16px] md:mx-[0]" : "bg-grey21 h-[106px] md:w-[calc(100%-72px)]"}`}
 		>
-			{isActive && (
+			{model.isActive && (
 				<div className="rounded-[50%] flex justify-center items-center w-[52px] h-[52px] bg-blue8">
 					<Image
 						src="/images/icons/user-badge-icon.svg"
@@ -40,40 +21,42 @@ export default function MemberCard({
 			<div className="flex flex-col justify-center gap-[8px]">
 				<div className=" flex justify-between">
 					<p className="text-[14px] leading-[14px] font-inconsolata text-grey4">
-						{doseTitle}
+						{model.doseTitle}
 					</p>{" "}
-					{doseValue && (
+					{model.doseValue && (
 						<p className="text-[14px] leading-[14px] text-blue font-inconsolata">
-							{doseValue}
+							{model.doseValue}
 						</p>
 					)}
 				</div>
 				<div className=" flex justify-between">
-					<p className="text-[24px] leading-[27px] font-hagerman">{name}</p>
+					<p className="text-[24px] leading-[27px] font-hagerman">
+						{model.name}
+					</p>
 					<div className="flex gap-[5px] items-center">
 						<span className="text-[20px] leading-[21px] font-bold font-inconsolata">
-							£{price.toFixed(2)}
+							£{model.price.toFixed(2)}
 						</span>{" "}
 						<p className="text-[12px] leading-[13px] font-bold font-inconsolata text-grey1">
-							(£{capsulePrice} / capsule)
+							(£{model.capsulePrice} / capsule)
 						</p>
 					</div>
 				</div>
 				<div className="flex justify-between">
 					<div className="flex gap-[5px] items-center text-[14px] leading-[14px] font-inconsolata text-grey4">
-						{discountTitle}
-						{forever && (
+						{model.discountTitle}
+						{model.forever && (
 							<span
-								className={`uppercase text-[14px] leading-[14px] font-inconsolata ${!isActive ? "text-blue font-[800]" : "text-grey6"}`}
+								className={`uppercase text-[14px] leading-[14px] font-inconsolata ${!model.isActive ? "text-blue font-[800]" : "text-grey6"}`}
 							>
 								Forever
 							</span>
 						)}
 					</div>
-					{spotsRemainds && (
+					{model.spotsRemainds && (
 						<p className="text-[14px] leading-[14px] font-inconsolata text-blue">
 							<span className="text-[14px] leading-[14px] font-inconsolata font-bold">
-								{spotsRemainds}{" "}
+								{model.spotsRemainds}{" "}
 							</span>
 							Founder Spots Remaining!
 						</p>
