+++ src/app/(main)/pre-order/components/OrderSummary.tsx
@@ -1,6 +1,7 @@
-import OrderCard from "@/components/cards/OrderCard";
 import { Separator } from "@radix-ui/react-separator";
-import Image from "next/image";
+
+import CorporationCardInfo from "@/components/cards/CorporationCardInfo";
+import OrderCard from "@/components/cards/OrderCard";
 
 export default function OrderSummary({
 	activeStep,
@@ -13,23 +14,7 @@ export default function OrderSummary({
 				</h1>
 			)}
 
-			<div className="grid gap-[8px]">
-				<p className="text-[20px] leading-[24px] font-inconsolata text-grey4">
-					KANEKA CORPRATION
-				</p>
-				<div className="text-[40px] leading-[48px] font-hagerman">
-					Coenzyme Q10 Ubiquinol Kaneka TM
-				</div>
-				<div className="flex items-center gap-[8px]">
-					<Image
-						src="/images/icons/link-icon.svg"
-						alt="security-card-icon"
-						width={24}
-						height={24}
-					/>
-					<p className="text-[14px] leading-[16px] text-grey6">100mg</p>
-				</div>
-			</div>
+			<CorporationCardInfo />
 
 			<Separator className="bg-grey13 h-[1px]" />
