+++ src/app/(main)/pre-order/components/PaymentDetails.tsx
@@ -1,9 +1,7 @@
 "use client";
-
 import { zodResolver } from "@hookform/resolvers/zod";
 import InputMask from "@mona-health/react-input-mask";
 import Image from "next/image";
-
 import { type SubmitHandler, useForm } from "react-hook-form";
 import { z } from "zod";
