+++ src/app/(main)/pre-order/components/ShareDiscounts.tsx
@@ -1,6 +1,7 @@
+import Link from "next/link";
+
 import ReferalCard from "@/components/cards/ReferalCard";
 import { Button } from "@/components/ui/button";
-import Link from "next/link";
 
 export default function ShareDiscounts() {
 	return (
