+++ src/app/(main)/products/[productID]/page.tsx
@@ -1,13 +1,14 @@
 "use client";
+import React from "react";
+
+import dynamic from "next/dynamic";
+import Image from "next/image";
+
+import { Separator } from "@radix-ui/react-separator";
+
 import ReferFriends from "@/components/ReferFriends";
 import OrderCard from "@/components/cards/OrderCard";
 import TotalCard from "@/components/cards/TotalCard";
-import {
-	Accordion,
-	AccordionContent,
-	AccordionItem,
-	AccordionTrigger,
-} from "@/components/ui/accordion";
 import {
 	Breadcrumb,
 	BreadcrumbItem,
@@ -25,153 +26,18 @@ import {
 	CarouselNext,
 	CarouselPrevious,
 } from "@/components/ui/carousel";
-import { Separator } from "@radix-ui/react-separator";
-import dynamic from "next/dynamic";
-import Image from "next/image";
-import React from "react";
+
+import ProductFaqs from "@/components/ProductFaqs";
+import CorporationBox from "../../pre-order/components/CorporationBox";
+import PreOrderInfo from "../../pre-order/components/PreOrderInfo";
+import Subscription from "../components/CapsuleBox";
 import ProfuctTable from "../components/ProductTable";
 import ReferalDiscounts from "../components/ReferalDiscounts";
-import Subscription from "../components/Subscription";
 import TeamPrice from "../components/TeamPrice";
 
 const StickyFooter = dynamic(() => import("@/components/StickyFooter"), {
 	ssr: false,
 });
-
-const images = [
-	{
-		id: 1,
-		src: "/images/icons/heart-pulse-icon.svg",
-		alt: "Checkmark icon",
-		title: "Heart Health",
-	},
-	{
-		id: 2,
-		src: "/images/icons/baby-icon.svg",
-		alt: "Baby icon",
-		title: "Fertility",
-	},
-	{
-		id: 3,
-		src: "/images/icons/energy-icon.svg",
-		alt: "Energy icon",
-		title: "Energy",
-	},
-	{
-		id: 4,
-		src: "/images/icons/dumbell-icon.svg",
-		alt: "Dumbell icon",
-		title: "Weight Training",
-	},
-	{
-		id: 5,
-		src: "/images/icons/hourglass-icon.svg",
-		alt: "Hourglass icon",
-		title: "Healthy Aging",
-	},
-	{
-		id: 6,
-		src: "/images/icons/tree-icon.svg",
-		alt: "Tree icon",
-		title: "Longevity",
-	},
-	{
-		id: 7,
-		src: "/images/icons/athletes-icon.svg",
-		alt: "Athletes icon",
-		title: "Athletes",
-	},
-	{
-		id: 8,
-		src: "/images/icons/brain-icon.svg",
-		alt: "Brain icon",
-		title: "Cognitive Function",
-	},
-	{
-		id: 9,
-		src: "/images/icons/water-drop-icon.svg",
-		alt: "Water drop icon",
-		title: "Skin Health",
-	},
-];
-
-const faqs = [
-	{
-		id: 1,
-		image: "/images/icons/cells-icon.svg",
-		imageAlt: "snowflake icon",
-		question: "Aging (40+)",
-		title: "Why Ubiquinol Matters",
-		description:
-			"As you age, your body produces less Ubiquinol, reducing your ability to create ATP, the energy your cells need to function. This leads to low energy, slower recovery, and reduced heart health. Ubiquinol helps replenish the active form of CoQ10, restoring cellular energy and protecting against oxidative stress.",
-		optionName: "Benefits",
-		options: [
-			"Increased energy levels ",
-			"Improved heart health",
-			"Slows down the aging process by reducing oxidative damage to cells",
-		],
-	},
-	{
-		id: 2,
-		image: "/images/icons/heart-pulse-icon.svg",
-		imageAlt: "heart pulse icon",
-		question: "Heart Health",
-		title: "",
-		description: "",
-		optionName: "",
-		options: [],
-	},
-	{
-		id: 3,
-		image: "/images/icons/dumbell-icon.svg",
-		imageAlt: "Dumbell icon",
-		question: "Training/Athletics",
-		title: "",
-		description: "",
-		optionName: "",
-		options: [],
-	},
-	{
-		id: 4,
-		image: "/images/icons/hourglass-icon.svg",
-		imageAlt: "Hourglass icon",
-		question: "Anti Aging",
-		title: "",
-		description: "",
-		optionName: "",
-		options: [],
-	},
-	{
-		id: 5,
-		image: "/images/icons/baby-icon.svg",
-		imageAlt: "Baby icon",
-		question: "Fertility",
-		title: "",
-		description: "",
-		optionName: "",
-		options: [],
-	},
-	{
-		id: 6,
-		image: "/images/icons/energy-icon.svg",
-		imageAlt: "Energy icon",
-		question: "Energy and Fatigue",
-		title: "",
-		description: "",
-		optionName: "",
-		options: [],
-	},
-	{
-		id: 7,
-		image: "/images/icons/brain-icon.svg",
-		imageAlt: "Brain icon",
-		question: "Brain Health",
-		title: "",
-		description: "",
-		optionName: "",
-		options: [],
-	},
-];
 
 const lastSectionItems = [
 	{
@@ -269,7 +135,7 @@ export default function ProductDetails({
 			setCurrent(api.selectedScrollSnap() + 1);
 		});
 	}, [api]);
-	const [loggedIn] = React.useState(true);
+	const [loggedIn] = React.useState(false);
 
 	return (
 		<div className="grid md:grid-cols-2 gap-[16px] min-h-screen container-width relative overflow-hidden">
@@ -329,111 +195,9 @@ export default function ProductDetails({
 					<TeamPrice members={175} />
 
 					<div className="grid gap-[60px] mt-[51px] md:mt-[0]">
-						<div className="grid gap-[24px]">
-							<div className="grid gap-[8px]">
-								<p className="text-[26px] leading-[39px] font-inconsolata font-bold">
-									Why Ubiquinol is Essential for Your Health
-								</p>
-								<p className="font-helvetica">
-									Ubiquinol powers cellular energy production and protects your
-									cells from oxidative stress, making it vital for staying
-									healthy, energetic, and vibrant as you age.S
-								</p>
-							</div>
+						<PreOrderInfo />
 
-							<Separator className="bg-black h-[1px]" />
-
-							<div className="grid gap-[8px]">
-								<p className="text-[20px] leading-[30px] font-inconsolata font-bold">
-									Boosts Cellular Energy
-								</p>
-								<p>
-									Ubiquinol is key to making ATP, the fuel your cells need to
-									function. Think of ATP as the battery that powers everything
-									from muscle movement to heart function. Without enough ATP,
-									your body slows down, causing fatigue and reduced performance.
-								</p>
-							</div>
-
-							<Separator className="bg-black h-[1px]" />
-
-							<div className="grid gap-[8px]">
-								<p className="text-[20px] leading-[30px] font-inconsolata font-bold">
-									Protects Against Free Radicals
-								</p>
-								<p>
-									Ubiquinol acts as a shield against free radicals—unstable
-									molecules that damage your cells. These molecules are produced
-									during normal activities (like exercising) and from external
-									factors (like pollution). Ubiquinol helps neutralize them,
-									preventing cell damage and slowing down the aging process,
-									especially in vital organs like your heart, brain, and
-									muscles.
-								</p>
-							</div>
-
-							<Separator className="bg-black h-[1px]" />
-
-							<div className="grid gap-[8px]">
-								<p className="text-[20px] leading-[30px] font-inconsolata font-bold">
-									Supports Healthy Aging
-								</p>
-								<p>
-									As you age, your body produces less Ubiquinol, which means
-									less ATP and fewer antioxidants. This leads to low energy,
-									slower recovery, and reduced heart health. Supplementing with
-									Ubiquinol helps restore these vital functions and keeps you
-									feeling energized and youthful.
-								</p>
-							</div>
-						</div>
-
-						<div>
-							{faqs.map((faq, index) => (
-								<Accordion key={faq.id} type="single" collapsible>
-									<Separator
-										className={`${index === faqs.length - 1 ? "" : ""} bg-black h-[1px]`}
-									/>
-									<AccordionItem value="products">
-										<AccordionTrigger>
-											<div className="grid gap-[16px] grid-cols-[32px_1fr] items-center text-[26px] leading-[27px] font-inconsolata font-bold text-black">
-												<Image
-													src={faq.image}
-													alt={faq.imageAlt}
-													width={32}
-													height={32}
-												/>{" "}
-												{faq.question}
-											</div>
-										</AccordionTrigger>
-										<AccordionContent>
-											<div className="mt-[-8px]">
-												<p className="text-[20px] leading-[30px] font-bold font-inconsolata mb-[8px]">
-													{faq.title}
-												</p>
-												<div className="text-[16px] leading-[24px] font-helvetica mb-[16px]">
-													{faq.description}
-												</div>
-												{faq.optionName && (
-													<div className="text-[16px] leading-[24px] font-bold font-helvetica italic mb-[8px]">
-														{faq.optionName}
-													</div>
-												)}
-												{faq.options && (
-													<ul className="list-disc list-inside">
-														{faq.options.map((option) => (
-															<li className="pl-[10px]" key={option}>
-																{option}
-															</li>
-														))}
-													</ul>
-												)}
-											</div>
-										</AccordionContent>
-									</AccordionItem>
-								</Accordion>
-							))}
-						</div>
+						<ProductFaqs />
 
 						<ProfuctTable />
 
@@ -477,46 +241,7 @@ export default function ProductDetails({
 					</BreadcrumbList>
 				</Breadcrumb>
 
-				<div className="grid gap-[8px]">
-					<p className="text-[20px] leading-[24px] md:font-[500] font-inconsolata md:text-grey4">
-						KANEKA CORPRATION
-					</p>
-					<div className="text-[24px] md:text-[40px] leading-[28px] md:leading-[48px] font-hagerman">
-						Coenzyme Q10 Ubiquinol Kaneka TM
-					</div>
-					<div className="flex items-center gap-[8px]">
-						<Image
-							src="/images/icons/link-icon.svg"
-							alt="security-card-icon"
-							width={24}
-							height={24}
-						/>
-						<p className="text-[14px] leading-[16px] text-grey6">100mg</p>
-					</div>
-				</div>
-
-				<div className="flex flex-col gap-[16px] font-helvetica">
-					<p>
-						Ubiquinol is the body-ready and active form of CoQ10 important for
-						virtually every cell in the body. It powers cellular energy
-						production and protects your cells from oxidative stress, making it
-						vital for staying healthy, energetic, and vibrant as you age.
-						<br />
-						<br />
-						Kaneka Corporation are the leading manufacturer based in Japan.
-					</p>
-					<div className="h-[56px] border-[1px] border-grey18 flex justify-center items-center gap-[16px]">
-						{images.map((image) => (
-							<Image
-								key={image.id}
-								src={image.src}
-								alt={image.alt}
-								width={24}
-								height={24}
-							/>
-						))}
-					</div>
-				</div>
+				<CorporationBox />
 
 				{!loggedIn && (
 					<div className="flex flex-col gap-[24px]">
