+++ src/app/(main)/products/components/ReferalDiscountCard.tsx
@@ -1,27 +1,19 @@
 import Image from "next/image";
 
-export default function ReferalDiscountCard({
-	name,
-	steps,
-	price,
-	priceOff,
-	isActive,
-}: {
-	name: string;
-	steps: number;
-	price: number;
-	priceOff: number;
-	isActive?: boolean;
-}) {
+import type { IReferalDiscountModel } from "@/utils/models/IReferalDiscountModel";
+
+export default function ReferalDiscountCard(
+	model: Readonly<IReferalDiscountModel>,
+) {
 	return (
 		<div
-			className={`grid grid-cols-[63px_1fr_auto] gap-[16px] p-[16px] md:px-[32px] items-center rounded-[8px] ${isActive ? "bg-white mx-[-16px] md:mx-[-18px] shadow-3 h-[127px]" : "bg-grey21 h-[95px]"}`}
+			className={`grid grid-cols-[63px_1fr_auto] gap-[16px] p-[16px] md:px-[32px] items-center rounded-[8px] ${model.isActive ? "bg-white mx-[-16px] md:mx-[-18px] shadow-3 h-[127px]" : "bg-grey21 h-[95px]"}`}
 		>
 			<div
-				className={`rounded-[50%] flex justify-center items-center w-[63px] h-[63px] ${isActive ? "bg-blue8" : " bg-grey20"}`}
+				className={`rounded-[50%] flex justify-center items-center w-[63px] h-[63px] ${model.isActive ? "bg-blue8" : " bg-grey20"}`}
 			>
 				<Image
-					src={`${isActive ? "/images/icons/add-user-icon.svg" : "/images/icons/lock-icon.svg"}`}
+					src={`${model.isActive ? "/images/icons/add-user-icon.svg" : "/images/icons/lock-icon.svg"}`}
 					alt="User profile group icon"
 					className="my-auto"
 					width={23}
@@ -30,25 +22,25 @@ export default function ReferalDiscountCard({
 			</div>
 			<div className="text-[14px] leading-[16px] font-helvetica flex flex-col justify-center gap-[8px]">
 				<p
-					className={`text-[20px] leading-[23px] uppercase font-hagerman ${isActive ? "text-black" : "text-grey6"}`}
+					className={`text-[20px] leading-[23px] uppercase font-hagerman ${model.isActive ? "text-black" : "text-grey6"}`}
 				>
-					{name}{" "}
+					{model.name}{" "}
 					<span
-						className={`text-[20px] leading-[23px] uppercase font-hagerman ${isActive ? "text-black" : "text-grey6"}`}
+						className={`text-[20px] leading-[23px] uppercase font-hagerman ${model.isActive ? "text-black" : "text-grey6"}`}
 					>
-						(1/{steps})
+						(1/{model.steps})
 					</span>
 				</p>
 				<p
-					className={`text-[20px] leading-[21px] font-inconsolata ${isActive ? "text-black font-[600]" : " text-grey15"}`}
+					className={`text-[20px] leading-[21px] font-inconsolata ${model.isActive ? "text-black font-[600]" : " text-grey15"}`}
 				>
-					£{priceOff} off Team Price
+					£{model.priceOff} off Team Price
 				</p>
 			</div>
 			<div
-				className={`font-inconsolata font-[800] flex items-center  ${isActive ? "text-blue text-[32px] leading-[33px]" : "text-grey15 text-[24px] leading-[25px]"}`}
+				className={`font-inconsolata font-[800] flex items-center  ${model.isActive ? "text-blue text-[32px] leading-[33px]" : "text-grey15 text-[24px] leading-[25px]"}`}
 			>
-				£{price.toFixed(2)}
+				£{model.price.toFixed(2)}
 			</div>
 		</div>
 	);
