+++ src/app/(main)/products/components/TeamPrice.tsx
@@ -1,6 +1,8 @@
-import type IUnlockTeamPriceMode from "@/utils/models/IUnlockTeamPriceMode";
-import { Separator } from "@radix-ui/react-separator";
 import Image from "next/image";
+
+import { Separator } from "@radix-ui/react-separator";
+
+import type IUnlockTeamPriceMode from "@/utils/models/IUnlockTeamPriceMode";
 
 const teamsItems = [
 	{
