+++ src/utils/models/ICorporationModel.ts
@@ -1,0 +1,6 @@
+export interface ICorporationModel {
+    id: number;
+    src: string;
+    alt: string;
+    title: string;
+}
