+++ src/utils/models/IFilterModel.ts
@@ -1,0 +1,4 @@
+export interface IFilterModel {
+    id: number;
+    value: string;
+}
