+++ src/utils/models/IReferalDiscountModel.ts
@@ -1,0 +1,7 @@
+export interface IReferalDiscountModel {
+    name: string;
+	steps: number;
+	price: number;
+	priceOff: number;
+	isActive?: boolean;
+}
